profit table
