+++ src/javascript/binary/websocket_pages/trade/beta/message.js
@@ -1,6 +1,7 @@
 var AssetIndexUI = require('../../resources/asset_index/asset_indexws.ui').AssetIndexUI;
 var MarketTimesUI = require('../../resources/market_times/market_timesws.ui').MarketTimesUI;
 var PortfolioWS = require('../../user/account/portfolio/portfolio.init').PortfolioWS;
+var ProfitTableWS = require('../../user/account/profit_table/profit_table.init').ProfitTableWS;
 
 /*
  * This Message object process the response from server and fire
