+++ src/javascript/binary/websocket_pages/user/account/portfolio/portfolio.init.js
@@ -5,7 +5,6 @@ var buildOauthApps = require('../../../../common_functions/get_app_details').bui
 var addTooltip = require('../../../../common_functions/get_app_details').addTooltip;
 var showTooltip = require('../../../../common_functions/get_app_details').showTooltip;
 var japanese_client = require('../../../../common_functions/country_base').japanese_client;
-var MBTradePage = require('../../../../websocket_pages/mb_trade/mb_tradepage').MBTradePage;
 var Portfolio = require('../portfolio').Portfolio;
 
 var PortfolioWS =  (function() {
@@ -201,7 +200,7 @@ var PortfolioWS =  (function() {
     };
 
     var onLoad = function() {
-        if (!TradePage.is_trading_page() && !TradePage_Beta.is_trading_page() && !MBTradePage.is_trading_page()) {
+        if (!/trading\.html/.test(window.location.pathname)) {
             BinarySocket.init({
                 onmessage: function(msg){
                     var response = JSON.parse(msg.data),
