+++ src/javascript/binary/websocket_pages/user/account/profit_table/profit_table.data.js
@@ -1,28 +1,4 @@
-var showLocalTimeOnHover = require('../../../../base/utility').showLocalTimeOnHover;
-var buildOauthApps = require('../../../../common_functions/get_app_details').buildOauthApps;
-var addTooltip = require('../../../../common_functions/get_app_details').addTooltip;
-
 var ProfitTableData = (function(){
-
-    function initSocket(){
-        BinarySocket.init({
-            onmessage: function(msg){
-                var response = JSON.parse(msg.data);
-
-                if (response) {
-                    var type = response.msg_type;
-                    if (type === 'profit_table'){
-                        ProfitTableWS.profitTableHandler(response);
-                        showLocalTimeOnHover('td.buy-date,td.sell-date');
-                    } else if (type === 'oauth_apps') {
-                        addTooltip(ProfitTableUI.setOauthApps(buildOauthApps(response.oauth_apps)));
-                    }
-                }
-            }
-        });
-        BinarySocket.send({'oauth_apps': 1});
-    }
-
     function getProfitTable(opts){
         var req = {profit_table: 1, description: 1};
         if(opts){
@@ -34,7 +10,6 @@ var ProfitTableData = (function(){
 
     return {
         getProfitTable: getProfitTable,
-        initSocket: initSocket
     };
 }());
