+++ src/javascript/binary/websocket_pages/user/account/profit_table/profit_table.ui.js
@@ -5,6 +5,7 @@ var format_money = require('../../../../common_functions/currency_to_symbol').fo
 var format_money_jp = require('../../../../common_functions/currency_to_symbol').format_money_jp;
 var showTooltip = require('../../../../common_functions/get_app_details').showTooltip;
 var japanese_client = require('../../../../common_functions/country_base').japanese_client;
+var ProfitTable = require('../profit_table').ProfitTable;
 
 var ProfitTableUI = (function(){
     "use strict";
