+++ src/javascript/binary/websocket_pages/websocket_pjax.js
@@ -10,6 +10,7 @@ var PasswordWS = require('./user/account/change_password').PasswordWS;
 var PaymentAgentTransferSocket = require('./user/account/payment_agent_transfer').PaymentAgentTransferSocket;
 var japanese_client = require('../common_functions/country_base').japanese_client;
 var PortfolioWS = require('./user/account/portfolio/portfolio.init').PortfolioWS;
+var ProfitTableWS = require('./user/account/profit_table/profit_table.init').ProfitTableWS;
 
 pjax_config_page_require_auth("user/profit_table", function(){
     return {
